% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BWM.R
\name{apply.BWM}
\alias{apply.BWM}
\title{Function for applying the Best-Worst Method}
\usage{
apply.BWM(
  criteria.lst,
  worst.criteria,
  best.criteria,
  best.criteria.preference,
  worst.criteria.preference
)
}
\arguments{
\item{criteria.lst}{list of criteria}

\item{worst.criteria}{the worst criteria}

\item{best.criteria}{the best criteria}

\item{best.criteria.preference}{the comparison of the best criteria to others}

\item{worst.criteria.preference}{the comparison of the worst criteria to others}
}
\value{
the result of BWM
}
\description{
Function for applying the Best-Worst Method
}
\examples{
criteria.lst <- c("C1", "C2", "C3")
worst.criteria <- "C1"
best.criteria <- "C3"
best.criteria.preference <- c(8, 2, 1)
worst.criteria.preference <- c(1, 5, 8)
apply.BWM(criteria.lst, worst.criteria, best.criteria, best.criteria.preference, worst.criteria.preference)
}
