% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCDA_methods.R
\name{apply.FAHP}
\alias{apply.FAHP}
\title{Apply fuzzy AHP on criteria comparison matrix}
\usage{
apply.FAHP(A)
}
\arguments{
\item{A}{the comparison matrix}
}
\value{
the fuzzy weights for each criteria
}
\description{
Apply fuzzy AHP on criteria comparison matrix
}
