% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PROMETHEE.R
\name{apply.PROMETHEE}
\alias{apply.PROMETHEE}
\title{Function for applying PROMOTHEE I or II}
\usage{
apply.PROMETHEE(A, weights, type = "II")
}
\arguments{
\item{A}{the comparison matrix with the row names indicating the alternatives and colnames
indicating the criteria.}

\item{weights}{the weights of criteria.}

\item{type}{can be either type 'I' or 'II'. It is set to 'II' by default}
}
\value{
the results of PROMOTHEE
}
\description{
Function for applying PROMOTHEE I or II
}
\examples{
A <- matrix(c(250, 200, 300, 275, 16, 16, 32, 32, 12, 8, 16, 8, 5, 3, 4, 2), nrow=4)
rownames(A)<-c("Mobile 1", "Mobile 2", "Mobile 3", "Mobile 4")
colnames(A)<-c("Price", "Memory", "Camera", "Looks")
weights <- c(0.35, 0.25, 0.25, 0.15)
apply.PROMETHEE(A, weights)
}
