% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCDA_methods.R
\name{apply.VIKOR}
\alias{apply.VIKOR}
\title{Function for applying VIKOR to data}
\usage{
apply.VIKOR(A, weights, nu = 0.5)
}
\arguments{
\item{A}{the comparison matrix}

\item{weights}{the weights of criteria}

\item{nu}{weight of the maximum utility strategy - set by default to 0.5}
}
\description{
Function for applying VIKOR to data
}
