% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCDA_methods.R
\name{find.entropy}
\alias{find.entropy}
\title{Find entropy of each criteria}
\usage{
find.entropy(A)
}
\arguments{
\item{A}{the matrix A with row names corresponding to alternatives and column
names corresponding to criteria}
}
\value{
the entropy value corresponding to each criteria
}
\description{
Find entropy of each criteria
}
