% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AHP.R
\name{find.weight}
\alias{find.weight}
\title{Finding the weights for each criteria given a pairwise comparison matrix A in the AHP method}
\usage{
find.weight(A)
}
\arguments{
\item{A}{the matrix containing information related to pairwise comparisons of
criteria}
}
\value{
a list containing the value of CI/RI and a vector containing the weights
of each criteria
}
\description{
Finding the weights for each criteria given a pairwise comparison matrix A in the AHP method
}
