% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCDA_methods.R
\name{plot.AHP.decision.tree}
\alias{plot.AHP.decision.tree}
\title{Plot decision tree}
\usage{
\method{plot}{AHP.decision.tree}(A, comparing.competitors)
}
\arguments{
\item{A}{the comparison matrix}

\item{comparing.competitors}{the list of matrices related to pairwise comparisons
of competitors for each criteria}
}
\value{
the decision tree plot
}
\description{
Plot decision tree
}
