% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_spider.R
\name{plot.spider}
\alias{plot.spider}
\title{Plot spider plot}
\usage{
\method{plot}{spider}(x, colors = palette("default"), ...)
}
\arguments{
\item{x}{the result of MCDA scores}

\item{colors}{the color scheme of choice}

\item{...}{further arguments passed to or from other methods}
}
\value{
the spider plot
}
\description{
Plot spider plot
}
