% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reading_CSV.R
\name{read.csv.AHP.matrices}
\alias{read.csv.AHP.matrices}
\title{Read csv file containing pairwise comparison matrices for applying AHP or ANP}
\usage{
read.csv.AHP.matrices(data)
}
\arguments{
\item{data}{the matrix containing information related to pairwise comparisons of
criteria}
}
\value{
a list containing a matrix A related to pairwise comparison of criteria
and a list containing multiple matrices related to pairwise comparisons of different
competitor products
}
\description{
Read csv file containing pairwise comparison matrices for applying AHP or ANP
}
\examples{
data <- read.csv(system.file("extdata", "AHP_input_file.csv",
 package = "RMCDA"), header=FALSE)
mat.lst <- read.csv.AHP.matrices(data)
}
