% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reading_CSV.R
\name{read.csv.SBWM.matrices}
\alias{read.csv.SBWM.matrices}
\title{Read csv file containing input to the stratified BWM method}
\usage{
read.csv.SBWM.matrices(data)
}
\arguments{
\item{data}{input of the csv file}
}
\value{
the inputs to the SBWM method
}
\description{
Read csv file containing input to the stratified BWM method
}
\examples{
data <- read.csv(system.file("extdata",
"stratified_BWM_case_study_I_example.csv",
 package = "RMCDA"), header = FALSE)
mat.lst <- read.csv.SBWM.matrices(data)
}
