% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reading_CSV.R
\name{read.csv.SMCDM.matrices}
\alias{read.csv.SMCDM.matrices}
\title{Read csv file containing pairwise comparison matrices for applying SMCDM}
\usage{
read.csv.SMCDM.matrices(data)
}
\arguments{
\item{data}{the matrix containing information related to pairwise comparisons of
criteria}
}
\value{
a list containing a matrix A related to pairwise comparison of criteria
and a list containing multiple matrices related to pairwise comparisons of different
competitor products
}
\description{
Read csv file containing pairwise comparison matrices for applying SMCDM
}
\examples{
data <- read.csv(system.file("extdata", "SMCDM_input.csv", package = "RMCDA"), header = FALSE)
mat.lst <- read.csv.SMCDM.matrices(data)
}
