% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCDA_methods.R
\name{read.matrices}
\alias{read.matrices}
\title{Read csv file containing pairwise comparison matrices}
\usage{
read.matrices(data)
}
\arguments{
\item{data}{the matrix containing information related to pairwise comparisons of
criteria}
}
\value{
a list containing a matrix A related to pairwise comparison of criteria
and a list containing multiple matrices related to pairwise comparisons of different
competitor products
}
\description{
Read csv file containing pairwise comparison matrices
}
