% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_spider.R
\name{spider.plot}
\alias{spider.plot}
\title{Plot spider plot}
\usage{
spider.plot(data, colors = palette("default"))
}
\arguments{
\item{data}{the result of MCDA scores}

\item{colors}{the color scheme of choice}
}
\value{
the spider plot
}
\description{
Plot spider plot
}
