% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCDA_methods.R
\name{apply.ANP}
\alias{apply.ANP}
\title{Apply Analytical Network Process (ANP) on data}
\usage{
apply.ANP(A, comparing.competitors, power)
}
\arguments{
\item{A}{the matrix containing information related to pairwise comparisons of
criteria}

\item{comparing.competitors}{the list of matrices related to pairwise comparisons
of competitors for each criteria}

\item{power}{the power value of the supermatrix}
}
\value{
the limiting super matrix
}
\description{
Apply Analytical Network Process (ANP) on data
}
