% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCDA_methods.R
\name{apply.CRITIC}
\alias{apply.CRITIC}
\title{Apply CRITIC on comparison matrix}
\usage{
apply.CRITIC(A)
}
\arguments{
\item{A}{the matrix A with row names corresponding to alternatives and column
names corresponding to criteria}
}
\value{
the weight percentages related to matrix A obtained through the CRITIC method
}
\description{
Apply CRITIC on comparison matrix
}
