% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ENTROPY.R
\name{apply.entropy}
\alias{apply.entropy}
\title{Find entropy of each criteria}
\usage{
apply.entropy(A)
}
\arguments{
\item{A}{the matrix A with row names corresponding to alternatives and column
names corresponding to criteria}
}
\value{
the entropy value corresponding to each criteria
}
\description{
Find entropy of each criteria
}
\examples{
A <- matrix(c(250, 200, 300, 275,
 225, 16, 16, 32,
  32, 16, 12, 8,
   16, 8, 16, 5,
    3, 4, 4, 2), nrow=5, ncol=4)
colnames(A)<-c("Price", "Storage space",
 "Camera", "Looks")
rownames(A)<-paste0("Mobile ", seq(1, 5, 1))
A[,"Price"] <- -A[,"Price"]
apply.entropy(A)
}
