% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AHP.R
\name{apply.FAHP}
\alias{apply.FAHP}
\title{Apply fuzzy AHP on criteria comparison matrix}
\usage{
apply.FAHP(A)
}
\arguments{
\item{A}{the comparison matrix}
}
\value{
the fuzzy weights for each criteria
}
\description{
Apply fuzzy AHP on criteria comparison matrix
}
\examples{
# example code
data <- read.csv(system.file("extdata", "AHP_input_file.csv", package = "RMCDA"), header=FALSE)
mat.lst <- read.csv.AHP.matrices(data)
mat.lst[[1]]->A
result <- apply.FAHP(A)
}
