% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RIM.R
\name{apply.RIM}
\alias{apply.RIM}
\title{Function to apply Reference Ideal Method (RIM)
Note:
function is rewritten from the MCDM package to match the formatting of the R RMCDA package
SOURCE: https://github.com/cran/MCDM/blob/master/R/RIM.R}
\usage{
apply.RIM(mat, weights, AB, CD)
}
\arguments{
\item{mat}{A matrix m x n containing the values of the m alternatives
for the n criteria.}

\item{weights}{A numeric vector of length n, containing the weights for the criteria.
The sum of the weights must be equal to 1.}

\item{AB}{A matrix (2 x n), where AB\link{1,} corresponds to the A extreme,
and AB\link{2,} corresponds to the B extreme of the domain (universe of discourse)
for each criterion.}

\item{CD}{A matrix (2 x n), where CD\link{1,} corresponds to the C extreme,
and CD\link{2,} corresponds to the D extreme of the ideal reference for each criterion.

Degenerate intervals:
\enumerate{
\item If AB\link{1,j} = CD\link{1,j}, then the interval \link{A, C} collapses to a point.
\itemize{
\item Any value x in this range is treated under a fallback rule:
\itemize{
\item If x = A = C, we set the normalized value to 1.
\item Otherwise, the normalized value is set to 0.
}
}
\item If CD\link{2,j} = AB\link{2,j}, then the interval \link{D, B} collapses to a point.
\itemize{
\item A similar fallback applies:
\itemize{
\item If x = D = B, we set the normalized value to 1.
\item Otherwise, the normalized value is set to 0.
}
}
}

These fallback rules ensure the function does not stop but, instead, issues a warning and assigns
a default. Adjust these defaults if your MCDM context requires different handling.}
}
\value{
A data frame containing:
\itemize{
\item Alternatives: The index of each alternative.
\item R: The R index (score) for each alternative.
\item Ranking: The ranking of the alternatives based on the R score.
}

Reference:
Cables, E.; Lamata, M.T.; Verdegay, J.L. (2016).
RIM-reference ideal method in multicriteria decision making.
Information Science, 337-338, 1-10.
}
\description{
The apply.RIM function implements the Reference Ideal Method (RIM) for
multi-criteria decision making (MCDM) problems, allowing for degenerate intervals,
i.e. cases where A == C or D == B.
}
\examples{

# Example decision matrix
mat <- matrix(
  c(30,40,25,27,45,0,
    9,0,0,15,2,1,
    3,5,2,3,3,1,
    3,2,3,3,3,2,
    2,2,1,4,1,2),
  nrow = 5, ncol = 6, byrow = TRUE
)

#Example weights vector (must sum to 1)
weights <- c(0.2262,0.2143,0.1786,0.1429,0.119,0.119)

#Example AB matrix
AB <- matrix(
  c(23,60,0,15,0,10,
    1,3,1,3,1,5),
  nrow = 2, ncol = 6, byrow = TRUE
)

#Example CD matrix
CD <- matrix(
  c(30,35,10,15,0,0,
    3,3,3,3,4,5),
  nrow = 2, ncol = 6, byrow = TRUE
)


apply.RIM(mat, weights, AB, CD)

}
