% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCDA_methods.R
\name{apply.TOPSIS}
\alias{apply.TOPSIS}
\title{Apply TOPSIS on matrix A with weight of criteria stored in vector w}
\usage{
apply.TOPSIS(A, w)
}
\arguments{
\item{A}{the matrix A with row names corresponding to alternatives and column
names corresponding to criteria}

\item{w}{the weight matrix corresponding to the weight of each criteria}
}
\value{
performance scores obtained through TOPSIS
}
\description{
Apply TOPSIS on matrix A with weight of criteria stored in vector w
}
